Move theming into its own file
