+++ src/react/constants.js
@@ -1,0 +1,1 @@
+export const CHANNEL = '__emotion__'
