+++ src/react/theming.js
@@ -1,0 +1,6 @@
+import { createTheming } from 'theming'
+import { CHANNEL } from './constants'
+
+const theming = createTheming(CHANNEL)
+const { withTheme, ThemeProvider } = theming
+export { CHANNEL, withTheme, ThemeProvider }
