+++ test/react.test.js
@@ -3,7 +3,8 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import { matcher, serializer } from '../jest-utils'
 import { css } from '../src/index'
-import styled, { ThemeProvider } from '../src/react'
+import styled from '../src/react'
+import { ThemeProvider } from '../src/react/theming'
 
 import { lighten, hiDPI, modularScale, borderWidth } from 'polished'
